+++ test/lsp/fixture/completion/template/childComponent/Child2143.vue
@@ -1,0 +1,12 @@
+<template>
+  
+</template>
+
+<script>
+export default {
+  /**
+   * my props documentation
+   */
+  props: ['foo']
+}
+</script>