+++ test/lsp/fixture/completion/template/childComponent/Parent2143.vue
@@ -1,0 +1,13 @@
+<template>
+  <child ></child>
+</template>
+
+<script>
+import Child from './Child2143.vue'
+
+export default {
+  components: {
+    Child
+  }
+}
+</script>