+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### 0.27.0
 
+- Improve cross-file completion when declaring simple props with `props: ['foo']`. #2143.
+- Completing child component should trigger props with `:` by default. #2140.
+- Space should trigger completion only in HTML mode. #2139.
+- Self-closing `<PascalCase>` component should get highlighted like JSX/TSX. #2136
+- Cross file template type checking - check that components are passed all declared props. #2135.
 - Linkify all vue/vue-router tags to their API doc. #2133.
 - Component Data - `type: 'boolean'` should trigger completion without `=""`. #2127.
 - Component Data doesn't work if it comes from devDependencies. #2132.